feat: 코드정리 dquote> - _pages 디렉토리 제거 dquote> - @wepl/ui 컴포넌트 위치 변경에 따른 설정 수정
