+++ apps/client/app/loading.tsx
@@ -1,8 +1,7 @@
-import LoadingView from '@/src/_pages/Loading';
 import React from 'react';
 
 const Loading = () => {
-  return <LoadingView />;
+  return <div>Loading...</div>;
 };
 
 export default Loading;
