+++ apps/client/app/onboarding/2.jpeg
