+++ apps/client/app/onboarding/page.tsx
@@ -1,15 +1,25 @@
-import { OnboardingView } from '@/src/_pages/Onboarding';
-import React from 'react';
+import * as React from 'react';
 
-const Page = () => {
+import { Carousel, CarouselContent, CarouselItem, CarouselNext, CarouselPrevious } from '@wepl/ui/Carousel.tsx';
+import Image1 from './1.jpeg';
+import Image2 from './2.jpeg';
+import Image from 'next/image';
+
+export default function OnboardingView() {
+  const imageArray = [Image1, Image2];
   return (
-    <div className="relative h-screen w-full flex flex-col">
-      <header className="self-end p-5">건너뛰기</header>
-      <section className="flex flex-col items-center justify-center text-center flex-grow">
-        <OnboardingView />
-      </section>
-    </div>
+    <Carousel className="w-full">
+      <CarouselContent>
+        {Array.from({ length: 2 }).map((_, index) => (
+          <CarouselItem key={index}>
+            <div className="p-1 flex justify-center items-center">
+              <Image src={imageArray[index]} alt="wepl" />
+            </div>
+          </CarouselItem>
+        ))}
+      </CarouselContent>
+      {/* <CarouselPrevious /> */}
+      {/* <CarouselNext /> */}
+    </Carousel>
   );
-};
-
-export default Page;
+}
