--- apps/client/src/_pages/Loading/index.tsx
@@ -1,5 +1,0 @@
-const LoadingView = () => {
-  return <div>Loading...</div>;
-};
-
-export default LoadingView;
