+++ packages/config-tailwind/tailwind.config.ts
@@ -7,7 +7,7 @@ const config: Config = {
     './components/**/*.{ts,tsx}',
     './app/**/*.{ts,tsx}',
     './src/**/*.{ts,tsx}',
-    '../../packages/ui/components/**/*.{ts,tsx}',
+    '../../packages/ui/src/*.{ts,tsx}',
   ],
   prefix: '',
   theme: {
