+++ packages/ui/components.json
@@ -11,7 +11,7 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "@ui/shadcn-ui",
+    "components": "@ui/src/",
     "utils": "@ui/lib/utils",
     "ui": "@ui/shadcn-ui"
   }
